hw-6

+++ controllers/auth/auth.controller.js
@@ -33,25 +33,25 @@ module.exports = {
         res.sendStatus(200);
     },
 
-    refreshToken: async (req, res) => {
-        const refresh_token = req.get(Authorization);
+    refreshToken: async (req, res, next) => {
+        try {
+            const refresh_token = req.get(Authorization);
 
-        await authService.deleteByParams({refresh_token});
+            const {userId} = await authService.getTokensByParams({refresh_token});
 
-        const {email, password} = req.body;
-        const user = await userService.getUserByParams({email});
+            if (!userId) {
+                return next(new ErrorHandler('NO USER', 404, 4041));
+            }
+            await authService.deleteByParams({refresh_token});
 
-        if (!user) {
-            return next(new ErrorHandler('NO USER', 404, 4041));
+            const tokens = tokinizer();
+
+            await authService.createTokenPair({userId, ...tokens});
+
+            res.json(tokens);
+        } catch (e) {
+            next(new ErrorHandler(e.message));
         }
-
-        await checkHashPassword(user.password, password);
-        
-        const tokens = tokinizer();
-
-        await authService.createTokenPair({...tokens, userId: user.id});
-
-        res.json(tokens);
-
     }
+
 };
